Remove extra constants file
