--- frontend/src/pages/ActivityReport/Pages/constants.js
@@ -1,74 +1,0 @@
-export const granteeParticipants = [
-  'CEO / CFO / Executive',
-  'Center Director / Site Director',
-  'Coach',
-  'Direct Service: Other',
-  'Family Service Worker / Case Manager',
-  'Fiscal Manager/Team',
-  'Governing Body / Tribal Council / Policy Council',
-  'Home Visitor',
-  'Manager / Coordinator / Specialist',
-  'Parent / Guardian',
-  'Program Director (HS / EHS)',
-  'Program Support / Administrative Assistant',
-  'Teacher / Infant-Toddler Caregiver',
-  'Volunteer',
-];
-
-export const nonGranteeParticipants = [
-  'Local/State Agency(ies)',
-  'HSCO',
-  'OCC Regional Office',
-  'OHS Regional Office',
-  'Regional Head Start Association',
-  'Regional TTA Team / Specialists',
-  'State Early Learning System',
-  'State Head Start Association',
-  'Other',
-];
-
-export const reasons = [
-  'Below Competitive Threshold (CLASS)',
-  'Below Quality Threshold (CLASS)',
-  'Change in Scope',
-  'Full Enrollment',
-  'New Grantee',
-  'New Director or Management',
-  'New Program Option',
-  'New Staff / Turnover',
-  'Ongoing Quality Improvement',
-  'Planning/Coordination (also TTA Plan Agreement)',
-  'School Readiness Goals',
-  'Monitoring | Area of Concern',
-  'Monitoring | Noncompliance',
-  'Monitoring | Deficiency',
-];
-
-export const otherUsers = [
-  'User 1',
-  'User 2',
-  'User 3',
-];
-
-export const programTypes = [
-  'program type 1',
-  'program type 2',
-  'program type 3',
-  'program type 4',
-  'program type 5',
-];
-
-export const targetPopulations = [
-  'Affected by Child Welfare Involvement',
-  'Affected by Disaster',
-  'Affected by Substance Use',
-  'Children with Disabilities',
-  'Children Experiencing Homelessness',
-  'Dual-Language Learners',
-  'Pregnant Women',
-];
-
-export const topics = [
-  'first',
-  'second',
-];
